refactoring to reduce nesting #23
