+++ polire/interpolate/__init__.py
@@ -1,0 +1,6 @@
+from .custom.custom import CustomInterpolator
+from .random.random import Random
+from .idw.idw import Idw
+from .spline.bspline import BSpline
+from .trend.trend import Trend
+from .spatial.spatial import SpatialAverage
