+++ polire/interpolate/trend/polynomials.py
