+++ polire/interpolate/trend/trend.py
