+++ usage.py
@@ -3,10 +3,8 @@ import seaborn as sns
 import matplotlib.pyplot as plt
 import numpy as np
 
-from polire.random.random import Random
-from polire.trend.trend import Trend
-from polire.spline.bspline import BSpline
-from polire.idw.idw import Idw
+from polire.interpolate import (Random, Trend, BSpline, Idw)
+
 # sample data
 X = [[0, 0], [0, 3], [3, 0], [3, 3]]
 y = [0, 1.5, 1.5, 3]
@@ -20,7 +18,7 @@ def test_grid():
         y_pred = r.predict_grid((0, 3), (0, 3))
         Z = y_pred
         sns.heatmap(Z)
-        plt.title(r.__class__)
+        plt.title(r)
         plt.show()
         plt.close()
 def test_point():
@@ -29,7 +27,7 @@ def test_point():
         r.fit(X,y)
         test_data = [[0,0],[0,3],[3,0],[3,3],[1,1],[1.5,1.5],[2,2],[2.5,2.5],[4,4]]
         y_pred = r.predict(np.array(test_data))
-        print(r.__class__)
+        print(r)
         print(y_pred)
 if __name__=='__main__':
     print("Testing Gridded Interpolation")
